+++ PCA9685.h
@@ -143,7 +143,7 @@ public:
     void enableExtClockLine();
 #endif
 
-    int getLastI2CError();
+    uint8_t getLastI2CError();
 
 private:
 #ifndef PCA9685_ENABLE_SOFTWARE_I2C
@@ -152,7 +152,7 @@ private:
     uint8_t _i2cAddress;        // Module's i2c address
     PCA9685_PhaseBalancer _phaseBalancer; // Phase balancer scheme to distribute load
     bool _isProxyAddresser;     // Instance is a proxy for sub addressing (disables certain functionality)
-    int _lastI2CError;          // Last i2c error
+    uint8_t _lastI2CError;      // Last i2c error
 
     void getPhaseCycle(int channel, uint16_t pwmAmount, uint16_t *phaseBegin, uint16_t *phaseEnd);
